Fixed error in transmission

+++ CircBuff/Core/Inc/circbuff.h
@@ -9,8 +9,8 @@
 
 typedef struct {
 	uint8_t buffer[CIRCBUFF_SIZE];
-	volatile uint16_t head;
-	volatile uint16_t tail;
+	volatile uint8_t head;
+	volatile uint8_t tail;
 	volatile bool full;
 
 } CircularBuffer;
